Refactor project modal handling with SearchParamsHandler for improved routing
